backend/requirements.txt
fastapi[standard]
pyjwt
passlib[bcrypt]
python-dotenv